backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg
bcrypt==5.0.0
click==8.3.2
colorama==0.4.6
email-validator==2.3.0
fastapi==0.135.3
greenlet==3.3.2
google-genai
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
openai
PyJWT==2.10.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.43.0
